feat: fungible token staking and transfers

+++ src/ghost-engine-frontend/src/hooks/useChunks.ts
@@ -15,7 +15,7 @@ export type FetchedChunk = {
   key: string;
   x: number;
   z: number;
-  data: Uint8Array | number[];
+  data: Uint16Array | number[];
   updatedAt: number;
 };
